requirements.txt
asgiref==3.11.1
dj-database-url==3.0.1
Django==4.2.28
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==26.0
pillow==11.3.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
typing_extensions==4.15.0
